+++ src/headless/main.cpp
@@ -46,7 +46,7 @@ static void print_rgo(ProvinceInstance const& province) {
 
 		auto const& employee_count_per_type_cache=rgo.get_employee_count_per_type_cache();
 		for (PopType const& pop_type : *employee_count_per_type_cache.get_keys()) {
-			const Pop::pop_size_t employees_of_type = employee_count_per_type_cache[pop_type];
+			const pop_size_t employees_of_type = employee_count_per_type_cache[pop_type];
 			if (employees_of_type > 0) {
 				text += StringUtils::append_string_views("\n\t\t", std::to_string(employees_of_type), " ", pop_type.get_identifier());
 			}
