+++ src/openvic-simulation/defines/PopsDefines.hpp
@@ -2,8 +2,8 @@
 
 #include "openvic-simulation/dataloader/NodeTools.hpp"
 #include "openvic-simulation/map/ProvinceInstance.hpp"
-#include "openvic-simulation/pop/Pop.hpp"
 #include "openvic-simulation/types/fixed_point/FixedPoint.hpp"
+#include "openvic-simulation/types/PopSize.hpp"
 #include "openvic-simulation/utility/Getters.hpp"
 
 namespace OpenVic {
@@ -71,8 +71,8 @@ namespace OpenVic {
 		fixed_point_t PROPERTY(nationalist_movement_mil_cap);
 		fixed_point_t PROPERTY(movement_support_uh_factor);
 		fixed_point_t PROPERTY(rebel_occupation_strength_bonus);
-		Pop::pop_size_t PROPERTY(large_population_limit);
-		Pop::pop_size_t PROPERTY(large_population_influence_penalty_chunk);
+		pop_size_t PROPERTY(large_population_limit);
+		pop_size_t PROPERTY(large_population_influence_penalty_chunk);
 
 		PopsDefines();
