+++ src/openvic-simulation/economy/production/Employee.cpp
@@ -2,7 +2,7 @@
 
 using namespace OpenVic;
 
-Employee::Employee(Pop& new_pop, const Pop::pop_size_t new_size)
+Employee::Employee(Pop& new_pop, const pop_size_t new_size)
 	: pop { new_pop },
 	size { new_size }
 	{}
