+++ src/openvic-simulation/economy/production/Employee.hpp
@@ -5,9 +5,9 @@
 namespace OpenVic {
 	struct Employee {
 	private:
-		Pop::pop_size_t PROPERTY_RW(size);
+		pop_size_t PROPERTY_RW(size);
 	public:
 		Pop& pop;
-		Employee(Pop& new_pop, const Pop::pop_size_t new_size);
+		Employee(Pop& new_pop, const pop_size_t new_size);
 	};
 }
