+++ src/openvic-simulation/economy/production/FactoryProducer.cpp
@@ -8,7 +8,7 @@ FactoryProducer::FactoryProducer(
 	fixed_point_t new_revenue_yesterday,
 	fixed_point_t new_output_quantity_yesterday,
 	fixed_point_t new_unsold_quantity_yesterday,
-	ordered_map<Pop*, Pop::pop_size_t>&& new_employees,
+	ordered_map<Pop*, pop_size_t>&& new_employees,
 	GoodDefinition::good_definition_map_t&& new_stockpile,
 	fixed_point_t new_budget,
 	fixed_point_t new_balance_yesterday,
