+++ src/openvic-simulation/economy/production/FactoryProducer.hpp
@@ -4,6 +4,7 @@
 
 #include "openvic-simulation/economy/GoodDefinition.hpp"
 #include "openvic-simulation/economy/production/ProductionType.hpp"
+#include "openvic-simulation/pop/Pop.hpp"
 #include "openvic-simulation/types/fixed_point/FixedPoint.hpp"
 #include "openvic-simulation/utility/Getters.hpp"
 
@@ -20,7 +21,7 @@ namespace OpenVic {
 		fixed_point_t PROPERTY(output_quantity_yesterday);
 		fixed_point_t PROPERTY(unsold_quantity_yesterday);
 		fixed_point_t PROPERTY(size_multiplier);
-		ordered_map<Pop*, Pop::pop_size_t> PROPERTY(employees);
+		ordered_map<Pop*, pop_size_t> PROPERTY(employees);
 		GoodDefinition::good_definition_map_t PROPERTY(stockpile);
 		fixed_point_t PROPERTY(budget);
 		fixed_point_t PROPERTY(balance_yesterday);
@@ -36,7 +37,7 @@ namespace OpenVic {
 		FactoryProducer(
 			ProductionType const& new_production_type, fixed_point_t new_size_multiplier, fixed_point_t new_revenue_yesterday,
 			fixed_point_t new_output_quantity_yesterday, fixed_point_t new_unsold_quantity_yesterday,
-			ordered_map<Pop*, Pop::pop_size_t>&& new_employees, GoodDefinition::good_definition_map_t&& new_stockpile,
+			ordered_map<Pop*, pop_size_t>&& new_employees, GoodDefinition::good_definition_map_t&& new_stockpile,
 			fixed_point_t new_budget, fixed_point_t new_balance_yesterday, fixed_point_t new_received_investments_yesterday,
 			fixed_point_t new_market_spendings_yesterday, fixed_point_t new_paychecks_yesterday, uint32_t new_unprofitable_days,
 			uint32_t new_subsidised_days, uint32_t new_days_without_input, uint8_t new_hiring_priority,
