+++ src/openvic-simulation/economy/production/ProductionType.cpp
@@ -20,7 +20,7 @@ ProductionType::ProductionType(
 	const std::optional<Job> new_owner,
 	std::vector<Job>&& new_jobs,
 	const template_type_t new_template_type,
-	const Pop::pop_size_t new_base_workforce_size,
+	const pop_size_t new_base_workforce_size,
 	GoodDefinition::good_definition_map_t&& new_input_goods,
 	GoodDefinition const& new_output_good,
 	const fixed_point_t new_base_output_quantity,
@@ -98,7 +98,7 @@ bool ProductionTypeManager::add_production_type(
 	std::optional<Job> owner,
 	std::vector<Job>&& jobs,
 	const ProductionType::template_type_t template_type,
-	const Pop::pop_size_t base_workforce_size,
+	const pop_size_t base_workforce_size,
 	GoodDefinition::good_definition_map_t&& input_goods,
 	GoodDefinition const* const output_good,
 	const fixed_point_t base_output_quantity,
@@ -256,7 +256,7 @@ bool ProductionTypeManager::load_production_types_file(
 			std::vector<Job> jobs;
 			ProductionType::template_type_t template_type { FACTORY };
 			GoodDefinition const* output_good = nullptr;
-			Pop::pop_size_t base_workforce_size = 0;
+			pop_size_t base_workforce_size = 0;
 			GoodDefinition::good_definition_map_t input_goods, maintenance_requirements;
 			fixed_point_t base_output_quantity = 0;
 			std::vector<ProductionType::bonus_t> bonuses;
