+++ src/openvic-simulation/map/ProvinceInstance.cpp
@@ -12,7 +12,6 @@
 #include "openvic-simulation/military/UnitInstanceGroup.hpp"
 #include "openvic-simulation/modifier/StaticModifierCache.hpp"
 #include "openvic-simulation/politics/Ideology.hpp"
-#include "openvic-simulation/pop/Pop.hpp"
 #include "openvic-simulation/utility/Logger.hpp"
 
 using namespace OpenVic;
