+++ src/openvic-simulation/types/PopSize.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <cstdint>
+
+namespace OpenVic {
+	using pop_size_t = int32_t;
+}
